Added attribution for GraphQLRequestContentReader
